move stream callbacks into api folder
