+++ nanome/api/streams/__init__.py
@@ -1,3 +1,3 @@
 from . import *
 from .stream import Stream
-from . import messages
+from . import messages, callbacks
